クラスを1ファイルにまとめてFinalを削除
