--- modules/__init__.py
@@ -1,1 +1,0 @@
-from .calendar_event import CalendarEvent
